backend/requirements.txt
fastapi
uvicorn
google-generativeai
langchain
langchain-community
langchain-core
langchain-groq
sentence-transformers
pypdf
faiss-cpu
pydantic
python-dotenv
python-multipart
numpy
